Moving code to proper file.
